requirements.txt
numpy
streamlit
pillow
scikit-learn
pandas